requirements.txt
openai==0.28.1
pandas==2.2.2
matplotlib==3.9.0
streamlit==1.29.0
pillow==10.1.0
numpy==1.26.2
